cdk-examples/pyproject.toml
[tool.poetry]
name = "cdk-examples"
version = "0.1.0"
description = ""
authors = ["Yaroslav Muravskyi <yaroslav.muravskyi@myarik.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-lambda-powertools = "^2.43.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
aws-cdk-lib = "^2.151.0"
ruff = "^0.5.7"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
"aws-cdk-aws-lambda-python-alpha" = "^2.151.0a0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
